Example integration test that takes device screenshots as it goes along
As the android-screenshot-lib is supported in android-maven-plugin,
getting a screenshot is simple - you just invoke 'poseForScreenshot()'
when you want one. The rest of the test is just setting up the activity
so it's ready to have it's screenshot taken - ie startActivitySync() and
then waiting a short period for the activity to be ready.

The screenshots will be placed in a 'screenshots' folder under your
target directory, for example:

morseflash/morseflash-instrumentation/target/screenshots

They're named sequentially by default (0000.png, 0001.png, etc), but you
can set an explicit name for the shot using 'poseForScreenshotNamed()'
if you like. In addition, for each android device used in the tests, an
animated-gif will be created showing all screenshots taken.

https://github.com/rtyley/android-screenshot-lib#readme

https://github.com/jayway/maven-android-plugin/pull/104